% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIONEERmetastaticTreatmentPLP.R
\docType{package}
\name{PIONEERmetastaticTreatmentPLP}
\alias{PIONEERmetastaticTreatmentPLP}
\alias{PIONEERmetastaticTreatmentPLP-package}
\title{PIONEERmetastaticTreatmentPLP}
\description{
PIONEERmetastaticTreatmentPLP
}
